+++ changelog.json
@@ -1,4 +1,7 @@
 {
+  "3.13.4" : {
+    "description" : "Page builder support. Color theme palette integration. Header options available for both posts and pages."
+  },
   "3.12.0" : {
     "description" : "Support php 8.1. Migrate theme to Sage version 10. Migrate Carbon fields to the new version. Remove Kernel Scripts. Migrate from webpack to bud."
   },
