+++ composer.json
@@ -72,5 +72,5 @@
       ]
     }
   },
-  "version": "3.14.3.6"
+  "version": "3.14.4.1"
 }
