+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "alps-wordpress-v3",
-  "version": "3.14.3.6",
+  "version": "3.14.4.1",
   "private": true,
   "browserslist": [
     "extends @wordpress/browserslist-config"
