+++ resources/styles/editor.css
@@ -4,10 +4,6 @@
 }
 
 .editor-styles-wrapper .wp-block {
-  max-width: min(calc(100vw - 8 * 25px), 610px);
-}
-
-.editor-styles-wrapper .wp-block {
   margin-left: auto;
   margin-right: auto;
 }
