+++ resources/views/layouts/app.blade.php
@@ -11,6 +11,18 @@
   if (is_singular()) {
       wp_enqueue_script('comment-reply');
   }
+
+  $remove_spacing = get_post_meta($post->ID, '_remove_spacing', true);
+  if ($remove_spacing){
+    $main_classes = '11 l-main';
+  }else{
+    $main_classes = '11 l-main u-spacing--double u-padding--double--bottom';
+  }
+
+  $template = get_post_meta($post->ID, '_wp_page_template', true);
+  if ($template == 'template-custom.blade.php'){
+    $theme_color_class.= ' template-custom';
+  }
 @endphp
 <html class="{{ $theme_color_class }}" {{ language_attributes() }}>
   @include('patterns.02-organisms.global.head')
@@ -19,7 +31,7 @@
       <div class="l-wrap__content l-content" role="document">
         @php do_action('get_header') @endphp
         @include('patterns.02-organisms.global.header')
-        <main class="11 l-main u-spacing--double u-padding--double--bottom" role="main">
+        <main class="{{ $main_classes }}" role="main">
           @yield('content')
         </main> <!-- /.l-main -->
         @php do_action('get_footer') @endphp
