+++ resources/views/patterns/02-organisms/sections/page-header.blade.php
@@ -35,6 +35,7 @@
   if (!is_home() && !$isPostsPage) {
     global $post;
     // PAGE FIELDS
+    $remove_header = get_post_meta($post->ID, $cf_.'remove_header', true);
     $hide_featured_image = get_post_meta($post->ID, $cf_.'hide_featured_image', true);
     $long_header_kicker = get_post_meta($post->ID, $cf_.'kicker', true);
     $long_header_title = get_post_meta($post->ID, $cf_.'display_title', true);
@@ -47,6 +48,12 @@
     if (!$header_background_image && empty($hide_featured_image)) {
       $header_background_image = get_post_thumbnail_id($post->ID);
     }
+  }
+
+  $page_header_classes = 'c-page-header c-page-header__long u-theme--background-color--dark  u-space--zero--top';
+  $feature_image_hero = get_post_meta($post->ID, $cf_.'featured_image_hero_layout', true);
+  if ($feature_image_hero == 'page-header'){
+    $page_header_classes = 'c-page-header c-page-header__long c-page-header__longer u-theme--background-color--dark  u-space--zero--top';
   }
 @endphp
 
@@ -77,22 +84,24 @@
   </style>
 @endif
 
-<header class="c-page-header c-page-header__long u-theme--background-color--dark  u-space--zero--top {{ $page_header_class }}">
-  <div class="c-page-header__long--inner l-grid l-grid--7-col {{ $page_header_inner_class }}">
-    <div class="c-page-header__content c-page-header__long__content l-grid-wrap l-grid-wrap--5-of-7 {{ $page_header_content_class }}">
-      @if ($long_header_kicker)
-        <span class="o-kicker u-color--white">{{ $long_header_kicker }}</span>
-      @endif
-      <h1 class="u-font--primary--xl u-color--white u-font-weight--bold">
-        {!! $long_header_title !!}
-      </h1>
+@if (empty($remove_header))
+  <header class="{{ $page_header_classes }} {{ $page_header_class }}">
+    <div class="c-page-header__long--inner l-grid l-grid--7-col {{ $page_header_inner_class }}">
+      <div class="c-page-header__content c-page-header__long__content l-grid-wrap l-grid-wrap--5-of-7 {{ $page_header_content_class }}">
+        @if ($long_header_kicker)
+          <span class="o-kicker u-color--white">{{ $long_header_kicker }}</span>
+        @endif
+        <h1 class="u-font--primary--xl u-color--white u-font-weight--bold">
+          {!! $long_header_title !!}
+        </h1>
+      </div>
     </div>
-  </div>
-</header>
-@if ($long_header_subtitle)
-  <div class="c-page-header__subtitle c-page-header__long__subtitle l-grid l-grid--7-col u-space--top--zero">
-    <div class="l-grid-wrap l-grid-wrap--5-of-7 u-shift--left--1-col--at-medium u-border--left u-font--secondary--m">
-      {{ $long_header_subtitle }}
+  </header>
+  @if ($long_header_subtitle)
+    <div class="c-page-header__subtitle c-page-header__long__subtitle l-grid l-grid--7-col u-space--top--zero">
+      <div class="l-grid-wrap l-grid-wrap--5-of-7 u-shift--left--1-col--at-medium u-border--left u-font--secondary--m">
+        {{ $long_header_subtitle }}
+      </div>
     </div>
-  </div>
+  @endif
 @endif
