+++ style.css
@@ -2,7 +2,7 @@
 Theme Name:         ALPS for WordPress
 Theme URI:          https://adventist.io/themes
 Description:        ALPS is the Adventist Living Pattern Library theme based off of the Sage WordPress theme.
-Version:            3.14.3.6
+Version:            3.14.4.1
 Author:             Seventh-day Adventist Church
 Author URI:         https://adventist.io/themes/worpdress
 Text Domain:        alps
